[WIP] Free tool handling from models.
This change removes ToolModel and ToolSelectionModel. Instead it adds
some methods to ProbeController and Probe to take over the
functionality. On the client side there's still ClientToolModel. But
now you can also access the tools directly without needing the model.

This is highly work in progress and doesn't even compile.
Current state:
- Server side seems to be done, it's completely untested though.
- Client side is almost completely missing and doesn't compile right
  now.

Things to do next:
- Port ClientToolModel to use the new API
- expose the new API via UiIntegration
- move ClientToolModel::widgetForId() to UiIntegration